Add support for add-list-members jobs in JobsProvider

+++ packages/webapp/providers/TwipsJobsProvider.tsx
@@ -3,7 +3,7 @@ import { useListState } from "@mantine/hooks";
 import { openConfirmModal } from "@mantine/modals";
 import { showNotification } from "@mantine/notifications";
 import { SupabaseClient } from "@supabase/supabase-js";
-import { Relation } from "@twips/common";
+import { JobName } from "@twips/common";
 import {
   createContext,
   ReactNode,
@@ -11,24 +11,22 @@ import {
   useEffect,
   useState,
 } from "react";
-import { useTwipsUser } from "./TwipsUserProvider";
 
 export type Job = {
   id: number;
+  name: JobName;
   createdAt: Date;
+  label: string;
   paused: boolean;
-  relation: Relation;
-  username: string;
-  updatedCount: number;
-  totalCount?: number;
   progress?: number;
+  totalCount?: number;
 };
 
 const TwipsJobsContext = createContext<{
   jobs: Job[];
   loading: boolean;
-  pauseJob: (id: number, paused: boolean) => Promise<void>;
-  deleteJob: (id: number) => Promise<void>;
+  pauseJob: (name: JobName, id: number, paused: boolean) => Promise<void>;
+  deleteJob: (name: JobName, id: number) => Promise<void>;
   refresh: () => void;
 }>({
   jobs: [],
@@ -44,17 +42,17 @@ interface TwipsJobsProviderProps {
 }
 
 const getJobs = async (supabase: SupabaseClient): Promise<Job[]> => {
-  const { data: jobs, error: selectJobsError } = await supabase
+  const { data: updateRelationJobsData } = await supabase
     .from("update_relation_job")
     .select(
       `id,created_at,paused,relation,updated_count,
         twitter_profile (username,followers_count,following_count)`
     )
     .eq("finished", false)
-    .order("created_at", { ascending: false });
-  if (selectJobsError) throw selectJobsError;
+    .order("created_at", { ascending: false })
+    .throwOnError();
 
-  return jobs.map((x) => {
+  const updateRelationJobs: Job[] = updateRelationJobsData.map((x) => {
     const updatedCount: number = x.updated_count;
     const totalCount: number =
       x.relation === "followers"
@@ -64,27 +62,52 @@ const getJobs = async (supabase: SupabaseClient): Promise<Job[]> => {
         ? // @ts-ignore
           x.twitter_profile.following_count
         : null;
-
     return {
       id: parseInt(x.id),
+      name: "update-relation",
+      label: `Fetch ${
+        x.relation == "blocking"
+          ? "blocklist"
+          : x.relation == "muting"
+          ? "mutelist"
+          : x.relation
+        // @ts-ignore
+      } of @${x.twitter_profile.username}`,
       createdAt: new Date(x.created_at),
       paused: x.paused,
-      relation: x.relation,
-      // @ts-ignore
-      username: x.twitter_profile.username,
-      updatedCount,
       totalCount,
       progress: totalCount ? (updatedCount / totalCount) * 100 : null,
     };
   });
+
+  const { data: addListMembersJobsData } = await supabase
+    .from("add_list_members_job")
+    .select(
+      `id,created_at,paused,member_ids,member_ids_added,
+        twitter_list (name)`
+    )
+    .eq("finished", false)
+    .order("created_at", { ascending: false })
+    .throwOnError();
+
+  const addListMembersJobs: Job[] = addListMembersJobsData.map((x: any) => {
+    return {
+      id: parseInt(x.id),
+      name: "add-list-members",
+      label: `Add ${x.member_ids.length} users to list "${x.twitter_list.name}`,
+      createdAt: new Date(x.created_at),
+      paused: x.paused,
+      progress: (x.member_ids_added.length / x.member_ids.length) * 100,
+    };
+  });
+
+  return [...updateRelationJobs, ...addListMembersJobs];
 };
 
 export const TwipsJobsProvider = ({
   supabase,
   children,
 }: TwipsJobsProviderProps) => {
-  const { user } = useTwipsUser();
-
   const [jobs, jobsHandlers] = useListState<Job>([]);
   const [loading, setLoading] = useState(false);
   const [randomFloat, setRandomFloat] = useState(Math.random());
@@ -112,13 +135,20 @@ export const TwipsJobsProvider = ({
   };
 
   // Handle pausing// unpausing of a job
-  const pauseJob = async (id: number, paused: boolean) => {
+  const pauseJob = async (name: JobName, id: number, paused: boolean) => {
     try {
-      await supabase
-        .from("update_relation_job")
-        .update({ paused })
-        .eq("id", id)
-        .throwOnError();
+      if (name == "update-relation")
+        await supabase
+          .from("update_relation_job")
+          .update({ paused })
+          .eq("id", id)
+          .throwOnError();
+      if (name == "add-list-members")
+        await supabase
+          .from("add_list_members_job")
+          .update({ paused })
+          .eq("id", id)
+          .throwOnError();
     } catch (error) {
       console.log(error);
       showNotification({
@@ -134,7 +164,7 @@ export const TwipsJobsProvider = ({
   };
 
   // Handle delete of a job
-  const deleteJob = async (id: number) => {
+  const deleteJob = async (name: JobName, id: number) => {
     openConfirmModal({
       title: "Are you sure you want to delete this job?",
       children: (
@@ -147,11 +177,18 @@ export const TwipsJobsProvider = ({
       onCancel: () => {},
       onConfirm: async () => {
         try {
-          await supabase
-            .from("update_relation_job")
-            .delete()
-            .eq("id", id)
-            .throwOnError();
+          if (name == "update-relation")
+            await supabase
+              .from("update_relation_job")
+              .delete()
+              .eq("id", id)
+              .throwOnError();
+          if (name == "add-list-members")
+            await supabase
+              .from("add_list_members_job")
+              .delete()
+              .eq("id", id)
+              .throwOnError();
         } catch (error) {
           console.log(error);
           showNotification({
@@ -172,7 +209,7 @@ export const TwipsJobsProvider = ({
     fetchJobs();
   }, [supabase, randomFloat]);
 
-  // Set event handler for checking progress
+  // Set event handler for checking progress of update-relation jobs
   useEffect(() => {
     supabase
       .channel(`public:update_relation_job`)
@@ -186,12 +223,40 @@ export const TwipsJobsProvider = ({
         (payload) => {
           // Update progress
           jobsHandlers.applyWhere(
-            (job) => job.id == payload.new.id,
+            (job) => job.id == payload.new.id && job.name == "update-relation",
             (job) => ({
               ...job,
               progress: job.totalCount
                 ? (payload.new.updated_count / job.totalCount) * 100
                 : null,
+            })
+          );
+        }
+      )
+      .subscribe();
+  }, []);
+
+  // Set event handler for checking progress of add-list-members jobs
+  useEffect(() => {
+    supabase
+      .channel(`public:add_list_members_job`)
+      .on(
+        "postgres_changes",
+        {
+          event: "UPDATE",
+          schema: "public",
+          table: "add_list_members_job",
+        },
+        (payload) => {
+          // Update progress
+          jobsHandlers.applyWhere(
+            (job) => job.id == payload.new.id && job.name == "add-list-members",
+            (job) => ({
+              ...job,
+              progress:
+                (payload.new.member_ids_added.length /
+                  payload.new.member_ids.length) *
+                100,
             })
           );
         }
